handle variable argument list separately for every log destination
fixes daemon crash whenever something is logged (perhaps only on amd64).
The value of ap after the return is indeterminate, so va_end has to
be called after vfprintf. It seems to be necessary to do this separately
for every command (e.g. vfprintf) in do_log that relies on the argument
list.

Conflicts:

	pgl/pgld/src/pgld.c