Fix printing IP for big-endian, change a lot of ints to 32-bit
Fix printing of IPs in big-endian (int2ip function)
Change name of assemble_ip to ip2int to match int2ip
Change a lot of int's to unsigned 32-bit int's
Fix issues if a range ends at 255.255.255.255.
Fix issues if a range is 0.0.0.0-255.255.255.255
Added listing the name of the file loading in the num entries loaded line.