updated and cleaned-up documentation for common pgl project
pgl is one project with modules (pgld, pglcmd and pgl-gui).
Cleaned-up for common documentation of the whole pgl project. Nearly all
documentation is in pgl/docs/ instead of in the separate modules
subfolders. (Started to) update the documentation.
Moved some old stuff to pgl-gui.qt-old/
Adapted the Makefiles to the new filenames (anticipating the renaming of
pgl-gui.qt to pgl-gui).