Removed the (non-default) feature to directly DROP and ACCEPT packets
removed the (non-default) feature to directly DROP and ACCEPT packets
(as in MoBlock 0.8), instead of using the (default) MARKing feature.

Reason:
ACCEPTing/DROPing packets directly was only needed for kernel/netfilter
not supporting MARKing. Probably it was also a bit more efficient than
MARKing.
But ACCEPTing packets directly breaks other iptables setups - this has
always been documented, but wasn't known to all users. So in order to
avoid misconfigurations I decided to drop this feature.
DROPing directly should not yield any drawbacks, but I decided to also
remove it in order to make the code simpler.