Added systemd timer units for daily blocklists update
Patch by Gilrain https://sourceforge.net/p/peerguardian/patches/3/
The mechanism is quite simple, consisting of 2 files:
a service dedicated to the command "pglcmd update" and
a timer calling the service at specific intervals.
The main pgl.service was updated to enable the timer at the same time
that pgl.service is enabled. This replicates the defaults state of the
cron script to make things simpler for users.
The rest of the diff is just makefile voodoo.

This only gets installed if pgl is compiled with --with-systemd.

Modified description (jre)