REPOSITORY: Remove pgosx/ folder (PeerGuardian OSX)
The PG OSX code is now in a separate repository of the peerguardian
project at sourceforge. That repository was created from a fork of this
one with:
git filter-branch --subdirectory-filter pgosx/

Get it with:
git clone git://git.code.sf.net/p/peerguardian/pgosx peerguardian-pgosx

This is part of a new repository layout with one project per repository.
This makes packaging for me easier. Of course a cross-platform PG is still
a goal, whether in this repository or just yet another one.